implement qjl
