+++ lib/quantization/Cargo.toml
@@ -32,6 +32,9 @@ parking_lot = { workspace = true }
 ordered-float = { workspace = true }
 arrayvec = "0.7.6"
 zerocopy = { workspace = true }
+nalgebra = "0.34"
+rand_chacha = "0.10.0"
+rand_distr = "0.6.0"
 
 [dev-dependencies]
 fs-err = { workspace = true, features = ["debug"] }
